+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the "Project Timer" extension will be documented in this file.
 
-## [Unreleased]
+## [0.3.0] - 2026-2-22
 ### Added
 - All cached data will be forced refresh 5 seconds after start up to avoid git extension not finishing scanning and cannot provide the correct Git remote URL.
 - The logs now will be written to VS Code `Output` panel.